Removed TypeDictionary and fixed PipelineGraph
I originally thought the problem was in how types were being looked up.
Turns out the PipelineGraph was already accounting for open/closed
generics. Just updated GetDefault instead, for an additional check if the
instance wasn't found and the requested type is generic, try the generic
type definition as the key.

The previous fix was waayy to invasive and caused other tests to fail.